Refactor directory structure for the enhanced CLI
