+++ pkg/cli/dev/cmd/list.go
@@ -4,9 +4,10 @@ import (
 	"fmt"
 	"os"
 
+	"github.com/tektoncd/results/pkg/cli/dev/flags"
+	"github.com/tektoncd/results/pkg/cli/dev/format"
+
 	"github.com/spf13/cobra"
-	"github.com/tektoncd/results/pkg/cli/flags"
-	"github.com/tektoncd/results/pkg/cli/format"
 	pb "github.com/tektoncd/results/proto/v1alpha2/results_go_proto"
 )
