+++ pkg/cli/dev/cmd/logs/logs.go
@@ -16,7 +16,7 @@ package logs
 
 import (
 	"github.com/spf13/cobra"
-	"github.com/tektoncd/results/pkg/cli/flags"
+	"github.com/tektoncd/results/pkg/cli/dev/flags"
 )
 
 // Command returns a cobra command for `logs` sub commands
